Move App to components
(All .svelte files to components)
