+++ src/main.js
@@ -1,4 +1,4 @@
-import App from './App.svelte';
+import App from 'components/App.svelte';
 import 'data/i18n.js';
 
 /**
